wireless-regdb: add country selection
Until now it has not been possible to remove countries from the regdb.
The packages has been extended, so that it is now possible to add or
omit individual countries in the regdb during package compile time.
If this option is not activated, then all countries in the regdb are taken
over as before.

If the package is updated to a new version. Then the script
`build_menuconfig` must be executed with the argument of the new db.txt to
update the country specification.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>