procd: add api wrapper
This commit adds the functions `procd_add_service_trigger` and
`procd_add_restart_service_trigger`. In order for the service to be
triggered by another service during a restart, only the function
procd_add_restart_service_trigger is needed.

Example:

service_trigger() {
	procd add_restart_service_trigger 'firewall' <command>
}

If the firewall is restarted, the specified <command> will be executed.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>