imagebuilder: add check if target is sourced from a feed
If a target is not upstream because it is still under development, it
can still be added to the openwrt via a feed and can be build. Although
the build is successful, no new images can be build with the new board,
because it cannot find them.

To fix this issue add a check to the imagebuilder and test if the target
is located in the feed directory.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>