usb: replace USB_OHCI_BIG_ENDIAN_* config with BIG_ENDIAN check
The kernel has a configuration check in the source whether the kernel
runs on big endian system. Until now, this had to be switched on explicitly
for the usb ohci in the kernel configuration for each target if the
system is 'BIG_ENDIAN'.

This commit checks in the 'package/kernel/linux/modules/usb.mk' whether a
'BIG_ENDIAN' system is to be compiled and if this is the case then the
required kernel configuration is switched on.

This change has the advantage that the needed ohci kernel option for
a 'BIG_ENDIAN' kernel no longer needs to be switched on explicitly.

USB_OHCI_BIG_ENDIAN_DESC:
https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git/tree/drivers/usb/host/ohci-platform.c?h=linux-6.6.y#n186
USB_OHCI_BIG_ENDIAN_MMIO:
https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git/tree/drivers/usb/host/ohci-platform.c?h=linux-6.6.y#n178

Signed-off-by: Florian Eckert <fe@dev.tdt.de>