treewide: remove INITRAMFS check for preinit_main hook
The 'preinit' script '/lib/preinit/70_initramfs_test' [1] checks whether
the system is running in an 'initramfs'. If this is the case, the loop [2]
in which the function is called is exited via a 'break' call. All further
'preinit_main' hooks are no longer processed. Therefore, the check whether
we are running in an initramfs is not necessary and are therefore removed.

[1] https://github.com/openwrt/openwrt/blob/master/package/base-files/files/lib/preinit/70_initramfs_test
[2] https://github.com/openwrt/openwrt/blob/master/package/base-files/files/lib/functions/preinit.sh#L57

Signed-off-by: Florian Eckert <fe@dev.tdt.de>